+++ tests/test_simstring.py
@@ -1,15 +1,37 @@
 import unittest
 
-from iamsystem import CacheFuzzyAlgos
-from iamsystem import Matcher
+from iamsystem.fuzzy.cache import CacheFuzzyAlgos
 from iamsystem.fuzzy.simstring import ESimStringMeasure
 from iamsystem.fuzzy.simstring import SimStringWrapper
+from iamsystem.matcher.matcher import Matcher
 
 
 class MatcherTest(unittest.TestCase):
     def setUp(self) -> None:
         self.matcher = Matcher()
         self.matcher.add_keywords(keywords=["paracetamol", "les"])
+
+    def test_init_algo_str(self):
+        """Check no error the measure is a string."""
+        fuzzy_ss = SimStringWrapper(
+            measure="DICE", words=["paracetamol"], threshold=0.2
+        )
+        syns = list(fuzzy_ss.get_syns_of_word("paracetamol"))
+        self.assertEqual(1, len(syns))
+
+    def test_init_algo_str_lower(self):
+        """Check no error the measure is lowercase."""
+        fuzzy_ss = SimStringWrapper(
+            measure="dice", words=["paracetamol"], threshold=0.2
+        )
+        syns = list(fuzzy_ss.get_syns_of_word("paracetamol"))
+        self.assertEqual(1, len(syns))
+
+    def test_init_algo_str_mispelled(self):
+        with (self.assertRaises(KeyError)):
+            SimStringWrapper(
+                measure="cosin", words=["paracetamol"], threshold=0.2
+            )
 
     def test_threshold_1(self):
         """Test threshold=1 is exact match"""